add resource class to manage ui

+++ app/src/main/java/com/basalam/basalamproduct/repository/ProductRepository.kt
@@ -7,27 +7,38 @@ import com.basalam.basalamproduct.db.ProductDatabase
 import com.basalam.basalamproduct.model.Product
 import com.basalam.basalamproduct.model.ProductResponse
 import com.basalam.basalamproduct.thread.ThreadExecutor
+import com.basalam.basalamproduct.util.Resource
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
 class ProductRepository(
     private val db: ProductDatabase,
-    private val threadExecutor: ThreadExecutor,
-    private val isLoading: MutableLiveData<Boolean>
+    private val threadExecutor: ThreadExecutor
 ) {
-    fun getProduct(query: String): LiveData<List<Product>> {
-        println("check for rotate ------------------------")
+    val productRes: MutableLiveData<Resource<LiveData<List<Product>>>> = MutableLiveData()
+    fun getProduct(query: String): MutableLiveData<Resource<LiveData<List<Product>>>> {
+        println("rotate log for test")
         threadExecutor.execute {
-            if (db.getProductDao().getProductSize() == 0)
-                isLoading.postValue(true)
+            productRes.postValue(Resource.Success(db.getProductDao().getAllProducts()))
+            if (db.getProductDao().getProductSize() == 0) {
+                productRes.postValue(Resource.Loading())
+            }
         }
+
         RetrofitInstance.api.getProduct(query).enqueue(
             object : Callback<ProductResponse> {
-                override fun onFailure(call: Call<ProductResponse>, t: Throwable) {
+                override fun onFailure(
+                    call: Call<ProductResponse>,
+                    t: Throwable
+                ) {
                     threadExecutor.execute {
-                        if (db.getProductDao().getProductSize() == 0)
-                            isLoading.postValue(false)
+                        productRes.postValue(
+                            Resource.Error(
+                                t.message!!,
+                                db.getProductDao().getAllProducts()
+                            )
+                        )
                     }
                 }
 
@@ -36,13 +47,34 @@ class ProductRepository(
                     response: Response<ProductResponse>
                 ) {
                     threadExecutor.execute {
-                        db.getProductDao().deleteAllProducts()
-                        db.getProductDao().insert(response.body()?.data?.productSearch?.products!!)
-                        isLoading.postValue(false)
+                        if (response.code() == 200) {
+                            println(response.body()?.data?.productSearch?.products!!.size.toString())
+                            if (db.getProductDao().getProductSize() == 0) {
+                                if (response.body()!!.data.productSearch.products.isEmpty()) {
+                                    productRes.postValue(Resource.Empty())
+                                } else {
+                                    productRes.postValue(
+                                        Resource.Success(
+                                            db.getProductDao().getAllProducts()
+                                        )
+                                    )
+                                }
+                            }
+                            db.getProductDao().deleteAllProducts()
+                            db.getProductDao()
+                                .insert(response.body()?.data?.productSearch?.products!!)
+                        } else {
+                            productRes.postValue(
+                                Resource.Error(
+                                    response.code().toString(),
+                                    db.getProductDao().getAllProducts()
+                                )
+                            )
+                        }
                     }
                 }
             }
         )
-        return db.getProductDao().getAllProducts()
+        return productRes
     }
 }
